Update to autofocus management
Autofocus has now been implemented as an inherit part of actioninput that is defult to true. It is now managed for pathgraph in colonyapp to avoid conflicting autofocusing actioninputs. When the player is not in the colony or when the handplacement is active (phases have been exposed) the pathgraph actioninput autofocus is disabled.